requirements.txt
scikit-image==0.17.2
pandas==1.0.3
numpy==1.18.1
Pillow==5.2.0